fix bug with deco metric & move design density metric
